requirements.txt
Pillow==9.5.0
grequests==0.7.0
python-dateutil==2.8.2
kdl-py==1.1.6
